Some more code cleanup.

+++ engine/include/physics/physics_world_settings.h
@@ -6,7 +6,7 @@ namespace Vulkyrie {
 
     struct PhysicsWorldSettings {
         public:
-            PhysicsWorldSettings(const std::string &name, const glm::vec3 &gravity = glm::vec3(0.0f, -9.81f, 0.0f), f64 frictionCoefficient = 0.5)
+            PhysicsWorldSettings(const std::string &name, const glm::vec3 &gravity = glm::vec3(0.0f, -9.81f, 0.0f), f32 frictionCoefficient = 0.5)
                 : Name(name)
                 , Gravity(gravity)
                 , FrictionCoefficient(frictionCoefficient) {
@@ -14,7 +14,7 @@ namespace Vulkyrie {
 
             std::string Name;
             glm::vec3 Gravity;
-            f64 FrictionCoefficient;
+            f32 FrictionCoefficient;
     };
 
 } // namespace Vulkyrie
